Updated docblocks and README; switched to PSR-4.
