+++ .gitignore
@@ -1,0 +1,3 @@
+composer.lock
+composer.phar
+/vendor/
