+++ composer.json
@@ -4,7 +4,7 @@
     "description": "Dependency injection containers interoperability",
     "license": "MIT",
     "autoload": {
-        "psr-0": {
+        "psr-4": {
             "Interop\\Container\\": "src/"
         }
     }
